requirements.txt
requests~=2.32.3
bs4~=0.0.2
aiogram~=2.23.1
telebot~=0.0.5
beautifulsoup4~=4.12.3
loguru
PyYAML
munch